Add variable and process variable after task query/pagination
The original activiti code had a hard coded limit of 20,000 entries when doing
a task query including process and task variables. This meant that the max results
that could be returned was 20,000 / (avg task var + avg process var) resulting in
empty pages if reqested beyond the limit.

While a 20,000 limit in active tasks may be ok it is less than desirable for
historic tasks. It also took many seconds to return a response due to the high memory
overhead of returning 20,000 entries.

The process now is
 1. Perform the query as if requested with task and process variables set to false
 2. If process or task variables are requested query for each task returned in step 1

This change will make more requests to the database but will ensure that pages past the
20,000 limit are not empty and performs better than in memory processing.